bug fix
the nodeLabel & linkList fields in the row template in the datagrid were
moving around and disappearing when the window was resized. turned out
that was because they had some random rules set in the geometry section
of property manager. cleared out the rules, allowing the layoutControl
to take care of everything, and now there's no problem.