draws node graphics and beginning of save .plat
I had to do a lot of experimentation in a different file to figure out
the graphics. Now some of that is implemented in the platyvue
application. Before implementing the graphical link part I want the file
save/open stuff to work so that I don't have to manually delete
thousands of link segments.