rewrote undo & redo so that they work with hide
replaced the algorithm. it's much better to just search for all of the
pointers attributed to the turn anywhere in the array and reverse them
than to search element by element through the array along a bread crumb
trail....particularly since hide doesn't leave the same trail as
undo/redo.