fixed a bug in link coordinate calculations
turns out there was a mistake in rotateCoord and translateCoord that
replaced empty lines with 0,0. it was masked because that extra
coordinate happened to make the extra lines overlap, but it was exposed
by the rise arrow pattern. now there are a few extra empty lines in the
calculated points that could be cleaned up, but they should be
irrelevant